LED blinking on each Picture, Tested on MIUI
My dev phone only have a white LED, no RGB. I cannot change the color.
To make it working I needed to long press on one of the notifications,
Click on TimeLapse, and enable Popup-Notification and LED-Light.
Then Restart the App.
Without Popup-Notification the LED does not turn on!

The LED is only turned on, if the screen is of.

#14

This is only an example, and not yet finished.